score_server/requirements.txt
click
flask
flask-login
flask-wtf
flask-sqlalchemy
gunicorn
requests
pytz
